requirements.txt
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-admin-tools==0.6.0
django-bootstrap-pagination==1.5.1
django-bootstrap3==6.2.2
django-etc==0.7.1
django-select-multiple-field==0.4.1
django-sitegate==0.10.2
django-sitemessage==0.7.0
django-toolbelt==0.0.1
django-widget-tweaks==1.4.1
gunicorn==19.3.0
nano==0.9.4
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
wsgiref==0.1.2